Refactor into bleu.py
